+++ libcontainer/cgroups/fscommon/utils.go
@@ -8,9 +8,17 @@ import (
 	"math"
 	"strconv"
 	"strings"
+
+	"github.com/opencontainers/runc/libcontainer/cgroups"
 )
 
-var ErrNotValidFormat = errors.New("line is not a valid key value format")
+var (
+	ErrNotValidFormat = errors.New("line is not a valid key value format")
+
+	OpenFile  = cgroups.OpenFile
+	ReadFile  = cgroups.ReadFile
+	WriteFile = cgroups.WriteFile
+)
 
 // ParseUint converts a string to an uint64 integer.
 // Negative values are returned at zero as, due to kernel bugs,
@@ -55,7 +63,7 @@ func ParseKeyValue(t string) (string, uint64, error) {
 // and returns a value of the specified key. ParseUint is used for value
 // conversion.
 func GetValueByKey(path, file, key string) (uint64, error) {
-	content, err := ReadFile(path, file)
+	content, err := cgroups.ReadFile(path, file)
 	if err != nil {
 		return 0, err
 	}
@@ -93,7 +101,7 @@ func GetCgroupParamUint(path, file string) (uint64, error) {
 // GetCgroupParamInt reads a single int64 value from specified cgroup file.
 // If the value read is "max", the math.MaxInt64 is returned.
 func GetCgroupParamInt(path, file string) (int64, error) {
-	contents, err := ReadFile(path, file)
+	contents, err := cgroups.ReadFile(path, file)
 	if err != nil {
 		return 0, err
 	}
@@ -111,7 +119,7 @@ func GetCgroupParamInt(path, file string) (int64, error) {
 
 // GetCgroupParamString reads a string from the specified cgroup file.
 func GetCgroupParamString(path, file string) (string, error) {
-	contents, err := ReadFile(path, file)
+	contents, err := cgroups.ReadFile(path, file)
 	if err != nil {
 		return "", err
 	}
